+++ src/solvers/dg/dg_2d.jl
@@ -110,35 +110,6 @@ function create_cache(mesh::TreeMesh{2}, equations, mortar_l2::LobattoLegendreMo
 end
 
 
-@inline function wrap_array(u_ode::AbstractVector, mesh::TreeMesh{2}, equations, dg::DG, cache)
-  @boundscheck begin
-    @assert length(u_ode) == nvariables(equations) * nnodes(dg)^ndims(mesh) * nelements(dg, cache)
-  end
-  # We would like to use
-  #   reshape(u_ode, (nvariables(equations), nnodes(dg), nnodes(dg), nelements(dg, cache)))
-  # but that results in
-  #   ERROR: LoadError: cannot resize array with shared data
-  # when we resize! `u_ode` during AMR.
-
-  # The following version is fast and allows us to `resize!(u_ode, ...)`.
-  # OBS! Remember to `GC.@preserve` temporaries such as copies of `u_ode`
-  #      and other stuff that is only used indirectly via `wrap_array` afterwards!
-  unsafe_wrap(Array{eltype(u_ode), ndims(mesh)+2}, pointer(u_ode),
-              (nvariables(equations), nnodes(dg), nnodes(dg), nelements(dg, cache)))
-end
-
-
-function compute_coefficients!(u, func, t, mesh::TreeMesh{2}, equations, dg::DG, cache)
-
-  @threaded for element in eachelement(dg, cache)
-    for j in eachnode(dg), i in eachnode(dg)
-      x_node = get_node_coords(cache.elements.node_coordinates, equations, dg, i, j, element)
-      u_node = func(x_node, t, equations)
-      set_node_vars!(u, u_node, equations, dg, i, j, element)
-    end
-  end
-end
-
 # TODO: Taal discuss/refactor timer, allowing users to pass a custom timer?
 
 function rhs!(du::AbstractArray{<:Any,4}, u, t,