+++ NEWS.md
@@ -15,6 +15,7 @@ for human readability.
   - multicomponent compressible Euler
   - acoustic perturbation equations
   - Lattice-Boltzmann equations 
+- New structured mesh type `StructuredMesh`
 
 #### Changed